requirements.txt
flask==2.0.1
twilio==7.0.0
matplotlib==3.4.3
pymongo==4.3.3
python-dotenv==0.19.0
razorpay==1.4.1 